+++ ai_server/human_detection_server.py
@@ -31,36 +31,6 @@ IMG_PATH = "received_image.JPG"
 #                       #
 #########################
 
-
-def test_human():
-    import cv2
-
-    # Load the pre-trained Haar Cascade classifier for human detection
-    human_cascade = cv2.CascadeClassifier(cv2.data.haarcascades + 'haarcascade_fullbody.xml')
-
-    # Load the image
-    image_path = 'temp2.jpg'
-    image = cv2.imread(image_path)
-
-    # # Convert the image to grayscale
-    # gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
-
-    # # Detect humans in the image
-    # humans = human_cascade.detectMultiScale(gray, scaleFactor=1.1, minNeighbors=5, minSize=(30, 30))
-
-    # Detect humans in the color image
-    humans = human_cascade.detectMultiScale(image, scaleFactor=1.1, minNeighbors=5, minSize=(30, 30))
-
-    # Draw rectangles around the detected humans
-    for (x, y, w, h) in humans:
-        cv2.rectangle(image, (x, y), (x + w, y + h), (255, 0, 0), 2)
-
-    # Display the image with detected humans
-    cv2.imshow('Humans Detected', image)
-    cv2.waitKey(0)
-    cv2.destroyAllWindows()
-
-
 def detect_humans(image_path):    
     # Initialize the Rekognition client
     client = boto3.client(
@@ -225,8 +195,7 @@ def start_tcp_server():
         server_socket.close()
 
 def main():
-    #start_tcp_server()
-    test_human()
+    start_tcp_server()
 
 if __name__ == "__main__":
     main()
